fix int to int64, and change QVector to std::vector in the implementation of AudioBuffer

+++ src/tests/simple/Audio2Test/buffer/AudioDataWrapper_p.h
@@ -13,8 +13,8 @@ public:
     AudioDataWrapper *q_ptr;
     float *const *data;
     int channelCount;
-    int sampleCount;
-    int startPos;
+    qint64 sampleCount;
+    qint64 startPos;
 
 };
